adgm-corporate-agent/requirements.txt
python-docx>=0.8.11
gradio>=3.0
openai>=0.27.0
faiss-cpu
tqdm
pdfplumber
spacy